// src/api/timepad/schema.d.ts
export interface components {
    schemas: {
// ...
        RegistrationOrderResponse: {
            id?: number;
            /** Format: date-time */
            created_at?: string;
            status?: components["schemas"]["OrderStatusResponse"];
            payment?: components["schemas"]["PaymentResponse"];
            tickets?: components["schemas"]["TicketResponse"][];
            answers?: {
                [key: string]: string;
            };
            promocodes?: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}